requirements.txt
Flask==2.0.1
Flask-HTTPAuth==4.2.0
Werkzeug==2.0.3
requests==2.25.1
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.20
setuptools>=71.1.0
Cython==0.29.37
numpy==1.23.0
pandas==1.5.0
Flask-WTF==0.15.1
gunicorn
